requirements.txt
bs4
requests
sqlalchemy
python-dateutil